+++ src/main/stalled_search_limit.h
@@ -7,7 +7,7 @@ namespace rows {
 
     class StalledSearchLimit : public operations_research::SearchLimit {
     public:
-        StalledSearchLimit(int time_limit_ms, operations_research::Solver *solver);
+        StalledSearchLimit(int64 time_limit_ms, operations_research::Solver *solver);
 
         bool Check() override;
 
@@ -28,7 +28,7 @@ namespace rows {
         bool found_first_solution_{false};
         int64 last_solution_update_{0};
 
-        int time_limit_ms_;
+        int64 time_limit_ms_;
     };
 }
